# packages/sdk/openapi/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AdottaUnGatto API",
    "version": "1.0.0",
    "description": "OpenAPI source for the v1 HTTP contract used by generated SDK types."
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/users/me/favorites/{listingId}": {
      "put": {
        "operationId": "addCurrentUserFavorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUserFavorite",
        "parameters": [
          {
            "$ref": "#/components/parameters/ListingId"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/users/me/linked-identities/{provider}/start": {
      "post": {
        "operationId": "startCurrentUserLinkedIdentity",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderAlias"
          }
        ],
        "responses": {
          "200": {
            "description": "Identity linking start information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartLinkedIdentityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/users/me/linked-identities/{provider}": {
      "delete": {
        "operationId": "deleteCurrentUserLinkedIdentity",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProviderAlias"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated linked identities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedIdentitiesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ListingId": {
        "name": "listingId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9]\\\\d*$"
        }
      },
      "ProviderAlias": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9][a-z0-9_-]{0,62}$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "UserFavorite": {
        "type": "object",
        "required": ["listingId", "addedAt"],
        "properties": {
          "listingId": {
            "type": "string"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FavoritesResponse": {
        "type": "object",
        "required": ["favorites"],
        "properties": {
          "favorites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserFavorite"
            }
          }
        }
      },
      "LinkedIdentity": {
        "type": "object",
        "required": [
          "provider",
          "providerSubject",
          "emailAtLink",
          "linkedAt",
          "lastSeenAt",
          "isPrimary"
        ],
        "properties": {
          "provider": {
            "type": "string"
          },
          "providerSubject": {
            "type": "string"
          },
          "emailAtLink": {
            "type": "string",
            "nullable": true
          },
          "linkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "isPrimary": {
            "type": "boolean"
          }
        }
      },
      "LinkedIdentitiesResponse": {
        "type": "object",
        "required": ["linkedIdentities"],
        "properties": {
          "linkedIdentities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkedIdentity"
            }
          }
        }
      },
      "StartLinkedIdentityResponse": {
        "type": "object",
        "required": ["provider", "redirectUrl"],
        "properties": {
          "provider": {
            "type": "string"
          },
          "redirectUrl": {
            "type": "string"
          }
        }
      },
...
    }
  }
}